Change return type of `Char.mutf8Length` from `Long` to `Int`
- The other `mutf8Length` overloads use `Long` because it's possible for them to exceed `Int.MAX_VALUE`
- `Char.mutf8Length` can only return `1`, `2`, or `3`, so there's no need for the extra range provided by a `Long`

+++ mew-tea-f8-common/src/commonMain/kotlin/mutf8Length.kt
@@ -16,7 +16,7 @@ import kotlin.jvm.JvmName
  *
  * @receiver The character to check the Modified UTF-8 length of.
  */
-val Char.mutf8Length: Long
+val Char.mutf8Length: Int
     get() = when (this.code.toShort().countLeadingZeroBits()) {
         // Characters in the range `'\u0001' .. '\u007F'` are encoded using `1` byte
         15, 14, 13, 12, 11, 10, 9 -> 1
